move all surface saving/reading/copying functions to general
