+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9surface.cpp
@@ -118,152 +118,4 @@ void surface_free(int id)
   delete enigma::surfaces[id];
 }
 
-int surface_getpixel(int id, int x, int y)
-{
-  get_surfacev(surface,id,-1);
-  if (x < 0) x = 0;
-  if (y < 0) y = 0;
-  if (x > surface.width || y > surface.height) return 0;
-  draw_batch_flush(batch_flush_deferred);
-
-  LPDIRECT3DSURFACE9 pBuffer = surface.surf, pRamBuffer;
-  enigma::surface_copy_to_ram(&pBuffer, &pRamBuffer);
-
-  D3DLOCKED_RECT rect;
-
-  pRamBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-  unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-  unsigned offset = y * rect.Pitch + x * 4;
-  int ret = bitmap[offset + 2] | (bitmap[offset + 1] << 8) | (bitmap[offset + 0] << 16);
-  pRamBuffer->UnlockRect();
-
-  pRamBuffer->Release();
-
-  return ret;
-}
-
-int surface_getpixel_ext(int id, int x, int y)
-{
-  get_surfacev(surface,id,-1);
-  if (x < 0) x = 0;
-  if (y < 0) y = 0;
-  if (x > surface.width || y > surface.height) return 0;
-  draw_batch_flush(batch_flush_deferred);
-
-  LPDIRECT3DSURFACE9 pBuffer = surface.surf, pRamBuffer;
-  enigma::surface_copy_to_ram(&pBuffer, &pRamBuffer);
-
-  D3DLOCKED_RECT rect;
-
-  pRamBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-  unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-  unsigned offset = y * rect.Pitch + x * 4;
-  int ret = bitmap[offset + 2] | (bitmap[offset + 1] << 8) | (bitmap[offset + 0] << 16) | (bitmap[offset + 3] << 24);
-  pRamBuffer->UnlockRect();
-
-  pRamBuffer->Release();
-
-  return ret;
-}
-
-int surface_getpixel_alpha(int id, int x, int y)
-{
-  get_surfacev(surface,id,-1);
-  if (x < 0) x = 0;
-  if (y < 0) y = 0;
-  if (x > surface.width || y > surface.height) return 0;
-  draw_batch_flush(batch_flush_deferred);
-
-  LPDIRECT3DSURFACE9 pBuffer = surface.surf, pRamBuffer;
-  enigma::surface_copy_to_ram(&pBuffer, &pRamBuffer);
-
-  D3DLOCKED_RECT rect;
-
-  pRamBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-  unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-  unsigned offset = y * rect.Pitch + x * 4;
-  int ret = bitmap[offset + 3];
-  pRamBuffer->UnlockRect();
-
-  pRamBuffer->Release();
-
-  return ret;
-}
-
-}
-
-//////////////////////////////////////SAVE TO FILE AND CTEATE SPRITE FUNCTIONS/////////
-//Fuck whoever did this to the spec
-#ifndef DX_BGR
-  #define DX_BGR 0x80E0
-#endif
-
-#include "Universal_System/estring.h"
-#include "Universal_System/image_formats.h"
-
-namespace enigma_user
-{
-
-int surface_save(int id, string filename)
-{
-  draw_batch_flush(batch_flush_deferred);
-
-  get_surfacev(surface,id,-1);
-  string ext = enigma::image_get_format(filename);
-
-  LPDIRECT3DSURFACE9 pDestBuffer;
-  D3DSURFACE_DESC desc;
-  surface.surf->GetDesc(&desc);
-
-  d3ddev->CreateOffscreenPlainSurface( desc.Width, desc.Height, desc.Format, D3DPOOL_SYSTEMMEM, &pDestBuffer, NULL );
-  d3ddev->GetRenderTargetData(surface.surf, pDestBuffer);
-
-  D3DLOCKED_RECT rect;
-
-  pDestBuffer->LockRect(&rect, NULL, D3DLOCK_READONLY);
-  unsigned char* bitmap = static_cast<unsigned char*>(rect.pBits);
-  pDestBuffer->UnlockRect();
-
-  int ret = enigma::image_save(filename, bitmap, desc.Width, desc.Height, desc.Width, desc.Height, false);
-
-  pDestBuffer->Release();
-
-  return ret;
-}
-
-int surface_save_part(int id, string filename, unsigned x, unsigned y, unsigned w, unsigned h)
-{
-  return 0; //TODO: implement
-}
-
-int background_create_from_surface(int id, int x, int y, int w, int h, bool removeback, bool smooth, bool preload)
-{
-  return 0; //TODO: implement
-}
-
-int sprite_create_from_surface(int id, int x, int y, int w, int h, bool removeback, bool smooth, bool preload, int xorig, int yorig)
-{
-  return 0; //TODO: implement
-}
-
-int sprite_create_from_surface(int id, int x, int y, int w, int h, bool removeback, bool smooth, int xorig, int yorig)
-{
-  return sprite_create_from_surface(id, x, y, w, h, removeback, smooth, true, xorig, yorig);
-}
-
-void sprite_add_from_surface(int ind, int id, int x, int y, int w, int h, bool removeback, bool smooth)
-{
-
-}
-
-void surface_copy_part(int destination, gs_scalar x, gs_scalar y, int source, int xs, int ys, int ws, int hs)
-{
-
-}
-
-void surface_copy(int destination, gs_scalar x, gs_scalar y, int source)
-{
-
-}
-
 }
